+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
-## [0.3.1] - 2023.xx.xx
+## [0.3.1] - 2023.11.08
 ### added
 - Add the log verbose for batch evaluation
 ### Changed
-- The lanelet vertices are now smoothed before constructing the curvlinear coordinate system; therefore, some of the test results vary a bit
+- The lanelet vertices are now extended and smoothed before constructing the curvlinear coordinate system; therefore, some of the test results vary a bit
 ### Fixed
 - The evasive maneuver simulator: now the initial time step of the simulated vehicle doesn't have to start from 0
 - The adaptive turning behaviors at intersections for the TTM simulation, i.e., the velocity needs to first satisfy the desired requirements based on the intersection's curvature